Refactor canvas font invalidation plumbing
Moves `subscribeToCanvasFontInvalidation` out of `canvasRenderHelpers` into `canvasSetup`, updates frame imports, and relocates the related test to `canvasSetup.test.ts`. It also removes `resolveCanvasFontFamily` and switches canvas label renderers/chrome code to resolve `--semiotic-font-family` directly via `resolveCSSColor`, keeping live and SSR text styling aligned. The network entry-graph gzip budget is raised from 156 KiB to 157 KiB to account for the added themed font handling overhead.
