+++ src/components/charts/xy/QuadrantChart.tsx
@@ -25,7 +25,7 @@ import { useXYPointStyle } from "../shared/useXYPointStyle"
 import { makeXYRuleContext, type StyleRule } from "../shared/styleRules"
 import { DEFAULT_QUADRANTS } from "./QuadrantChart.defaults"
 import { getMinMax } from "../shared/minMax"
-import { resolveCanvasFontFamily } from "../../stream/renderers/canvasRenderHelpers"
+import { resolveCSSColor } from "../../stream/renderers/resolveCSSColor"
 
 registerXYPlugin(scatterXYPlugin)
 
@@ -506,7 +506,8 @@ export const QuadrantChart = forwardRef(function QuadrantChart<TDatum extends Da
       const _cx = Math.max(0, Math.min(w, xC))
       const _cy = Math.max(0, Math.min(h, yC))
 
-      ctx.font = `600 ${quadrantLabelSize}px ${resolveCanvasFontFamily(ctx)}`
+      const fontFamily = resolveCSSColor(ctx, "var(--semiotic-font-family, sans-serif)") || "sans-serif"
+      ctx.font = `600 ${quadrantLabelSize}px ${fontFamily}`
       ctx.globalAlpha = 0.5
 
       const padding = 8
