+++ src/components/stream/StreamOrdinalFrame.tsx
@@ -41,7 +41,6 @@ import { StalenessBadge } from "./StalenessBadge"
 import { OrdinalSVGOverlay, OrdinalSVGUnderlay } from "./OrdinalSVGOverlay"
 import { resolveAnnotationAccessor, buildEnrichAnnotationData } from "./annotationAccessorResolver"
 import { OrdinalBrushOverlayLazy } from "./OrdinalBrushOverlayLazy"
-import { subscribeToCanvasFontInvalidation } from "./renderers/canvasRenderHelpers"
 import { isServerEnvironment } from "./isServerEnvironment"
 import { useHydration, useWasHydratingFromSSR } from "./useHydration"
 import { useStableShallow } from "./useStableShallow"
@@ -58,7 +57,7 @@ import { shouldHandleFramePointer } from "./frameCursorInteraction"
 import { rehitOrdinalFrameCursor } from "./ordinalFrameCursorInteraction"
 
 // Canvas setup / hover
-import { getDevicePixelRatio } from "./canvasSetup"
+import { getDevicePixelRatio, subscribeToCanvasFontInvalidation } from "./canvasSetup"
 import type { HoverPointerCoords } from "./hoverUtils"
 import { useLegendCategoryEmission } from "./useLegendCategoryEmission"
 import { resolveFrameGraphics } from "./frameGraphics"
