+++ src/components/stream/StreamXYFrame.tsx
@@ -31,7 +31,6 @@ import { isServerEnvironment } from "./isServerEnvironment"
 import { useHydration, useWasHydratingFromSSR } from "./useHydration"
 import { useStableShallow } from "./useStableShallow"
 import { paintCanvasBackground, resolveCanvasBackground } from "./canvasBackground"
-import { subscribeToCanvasFontInvalidation } from "./renderers/canvasRenderHelpers"
 import { needsInteractionCanvasPaint } from "./paintNeeds"
 import { createFrameThemeColorCache, LIGHT_FRAME_THEME } from "./frameThemeColors"
 
@@ -44,7 +43,7 @@ import { CanvasFrameBackground, useFrameCanvasHost } from "./useCanvasFrameHost"
 import { refreshIdlePulse } from "./pulseFrameRefresh"
 import { resolveFrameGraphics } from "./frameGraphics"
 
-import { prepareCanvas, getDevicePixelRatio, syncCanvasSize } from "./canvasSetup"
+import { prepareCanvas, getDevicePixelRatio, syncCanvasSize, subscribeToCanvasFontInvalidation } from "./canvasSetup"
 import { buildHoverData, getPointerHitRadius, type HoverPointerCoords } from "./hoverUtils"
 import { useLegendCategoryEmission } from "./useLegendCategoryEmission"
 import { filterSparseArray } from "../charts/shared/sparseArray"
