+++ src/components/stream/canvasSetup.ts
@@ -12,6 +12,18 @@ export interface CanvasMargin {
   left: number
 }
 
+/** Repaint canvas text after an asynchronously loaded web font becomes usable. */
+export function subscribeToCanvasFontInvalidation(
+  listener: () => void
+): () => void {
+  const fonts = typeof document === "undefined" ? undefined : document.fonts
+  if (!fonts) return () => {}
+  fonts.addEventListener("loadingdone", listener)
+  return () => {
+    fonts.removeEventListener("loadingdone", listener)
+  }
+}
+
 /**
  * Keep a canvas's backing store and CSS box aligned with the frame size.
  * This is separate from context preparation so an idle interaction layer can
