+++ src/components/stream/renderers/canvasRenderHelpers.test.ts
@@ -3,7 +3,6 @@ import {
   resolveCurveFactory,
   resolveCanvasFill,
   resolveCanvasPaint,
-  subscribeToCanvasFontInvalidation,
   buildLinearFillGradient,
   paintNetworkFill,
   paintNetworkStroke,
@@ -15,43 +14,6 @@ describe("canvasRenderHelpers", () => {
   let ctx: CanvasRenderingContext2D
   beforeEach(() => {
     ctx = createMockCanvasContext() as object as CanvasRenderingContext2D
-  })
-
-  it("notifies canvas subscribers when web fonts finish loading and cleans up", () => {
-    const listeners = new Set<EventListenerOrEventListenerObject>()
-    const fontSet = {
-      addEventListener: vi.fn((type: string, listener: EventListenerOrEventListenerObject) => {
-        if (type === "loadingdone") listeners.add(listener)
-      }),
-      removeEventListener: vi.fn((type: string, listener: EventListenerOrEventListenerObject) => {
-        if (type === "loadingdone") listeners.delete(listener)
-      })
-    }
-    const originalFonts = Object.getOwnPropertyDescriptor(document, "fonts")
-    Object.defineProperty(document, "fonts", { configurable: true, value: fontSet })
-    const listener = vi.fn()
-    const unsubscribe = subscribeToCanvasFontInvalidation(listener)
-
-    try {
-      for (const registered of listeners) {
-        if (typeof registered === "function") registered(new Event("loadingdone"))
-        else registered.handleEvent(new Event("loadingdone"))
-      }
-      expect(listener).toHaveBeenCalledTimes(1)
-      unsubscribe()
-      expect(listeners.size).toBe(0)
-      expect(fontSet.removeEventListener).toHaveBeenCalledWith(
-        "loadingdone",
-        expect.any(Function)
-      )
-    } finally {
-      unsubscribe()
-      Object.defineProperty(
-        document,
-        "fonts",
-        originalFonts || { configurable: true, value: undefined }
-      )
-    }
   })
 
   describe("resolveCurveFactory", () => {
